fix create dir and rename dir

See merge request itway/nft-collection-node-1!73

+++ src/services/deployTrueNft.service.ts
@@ -33,7 +33,9 @@ export class DeployTrueNftService {
         let address = "0";
         try {
             address = await this.deployRootNft(rootNftAccount, indexAccount, dataAccount, collection, commissionAuthorGenerator);
-            fs.renameSync(pathWithContracts, path.join(globals.RESULT_COLLECTION, address.slice(2)))
+            if (!fs.existsSync(path.join(globals.RESULT_COLLECTION, address.slice(2)))) {
+                fs.renameSync(pathWithContracts, path.join(globals.RESULT_COLLECTION, address.slice(2)))
+            }
             console.log("RootNft address: " + await rootNftAccount.getAddress());
             await this.deployBasis(rootNftAccount, indexBasisAccount);
         } catch(err) {
